Add README for skipping the DB entirely
Add a spreadsheet folder.
For now, just has a readme with a functional specification for doing all
the reporting from within the spreadsheet rather than using a database.
Most normal people can use a spreadsheet but do not have the time,
expertise, or inclincation to stand up a DB server.